TASK-56 add domain module
Add domain module to root project, move common entities from core module to domain module.
