--- core/src/main/kotlin/ru/surf/core/entity/QuestionType.kt
@@ -1,25 +1,0 @@
-package ru.surf.core.entity
-
-import ru.surf.core.entity.base.UUIDBasedEntity
-import java.util.*
-import javax.persistence.*
-
-@Table(name = "question_types")
-@Entity
-class QuestionType(
-
-    @Id
-    @Column(name = "id")
-    override val id: UUID = UUID.randomUUID(),
-
-    @Column(name = "type")
-    var type: String = "",
-
-    ) : UUIDBasedEntity(id) {
-
-    @Override
-    override fun toString(): String {
-        return this::class.simpleName + "(id = $id , type = $type )"
-    }
-
-}
