+++ domain/src/main/kotlin/ru/surf/core/entity/Answer.kt
@@ -1,0 +1,29 @@
+package ru.surf.core.entity
+
+import ru.surf.core.entity.base.UUIDBasedEntity
+import java.util.*
+import javax.persistence.Column
+import javax.persistence.Entity
+import javax.persistence.Id
+import javax.persistence.Table
+
+
+@Table(name = "answers")
+@Entity
+class Answer(
+
+        @Id
+        @Column(name = "id")
+        override val id: UUID = UUID.randomUUID(),
+
+        @Column(name = "answer")
+        var text: String = "",
+
+        ) : UUIDBasedEntity(id) {
+
+    @Override
+    override fun toString(): String {
+        return this::class.simpleName + "(id = $id , text = $text )"
+    }
+
+}
