+++ domain/src/main/kotlin/ru/surf/core/entity/EventType.kt
@@ -1,0 +1,28 @@
+package ru.surf.core.entity
+
+import ru.surf.core.entity.base.UUIDBasedEntity
+import java.util.*
+import javax.persistence.Column
+import javax.persistence.Entity
+import javax.persistence.Id
+import javax.persistence.Table
+
+@Table(name = "event_types")
+@Entity
+class EventType(
+
+        @Id
+        @Column(name = "id")
+        override val id: UUID = UUID.randomUUID(),
+
+        @Column(name = "type")
+        var description: String = "",
+
+        ) : UUIDBasedEntity(id) {
+
+    @Override
+    override fun toString(): String {
+        return this::class.simpleName + "(id = $id , description = $description )"
+    }
+
+}
