+++ domain/src/main/kotlin/ru/surf/core/entity/Trainee.kt
@@ -1,0 +1,44 @@
+package ru.surf.core.entity
+
+import ru.surf.core.entity.base.UUIDBasedEntity
+import java.util.*
+import javax.persistence.*
+
+@Table(name = "trainees")
+@Entity
+class Trainee(
+
+        @Id
+        @Column(name = "id")
+        override val id: UUID = UUID.randomUUID(),
+
+        @Column(name = "score")
+        var score: Int? = null,
+
+        @Column(name = "is_active")
+        var isActive: Boolean = false,
+
+        @ManyToOne(cascade = [CascadeType.REFRESH], fetch = FetchType.LAZY)
+        @JoinColumn(name = "candidate_id")
+        var candidate: Candidate = Candidate(),
+
+        @ManyToOne(cascade = [CascadeType.REFRESH], fetch = FetchType.LAZY)
+        @JoinColumn(name = "event_id")
+        var event: Event = Event(),
+
+        @ManyToOne(cascade = [CascadeType.REFRESH], fetch = FetchType.LAZY)
+        @JoinColumn(name = "team_id")
+        val team: Team? = null,
+
+        @OneToOne(cascade = [CascadeType.REFRESH], fetch = FetchType.LAZY)
+        @JoinColumn(name = "account_id")
+        var account: Account = Account(),
+
+        ) : UUIDBasedEntity(id) {
+
+    @Override
+    override fun toString(): String {
+        return this::class.simpleName + "(id = $id , score = $score , isActive = $isActive )"
+    }
+
+}
