+++ domain/src/main/kotlin/ru/surf/core/entity/base/BaseEntity.kt
