+++ domain/src/main/kotlin/ru/surf/core/entity/base/UUIDBasedEntity.kt
@@ -8,7 +8,7 @@ import javax.persistence.MappedSuperclass
 @MappedSuperclass
 abstract class UUIDBasedEntity(
 
-    @Id
-    @Column(columnDefinition = "uuid")
-    override val id: UUID
+        @Id
+        @Column(columnDefinition = "uuid")
+        override val id: UUID
 ) : BaseEntity<UUID>()
